Add file to releases
Due to the upload limitation, Lileilei scene has been upload into releases tagged HostFile1.0
This time, Lileilei Scene has been added in button 'Level 3' on menu1 and 'Level 1' on menu 2.
The purpose of this scene is to provide gamer an immersive feeling. The user could feel the gloomy atmosphere. As they come from the menu the scene the game will not begin until the user finish the dialogs of the introduction which gives the user a basic understanding of the game. Next, player can hear a jump sound and the game timer will begin along with some creepy music.
You can choose pause this game when playing it. Timer will stop during that time. And you may see game over pop up window or win window. Then you can choose go another page or play again.
The user will need to press space to finish the dialogs and continue enjoy the game.
The user can use key board control to choose player’s direction.
The user needs to use mouse to change the view angle.
The user needs user the weapon hit the enemy.
The user needs to choose the action after finish gaming.
Other interactions are not mandatory but if the user key down esc the game will pause. User needs to choose where they want to go on that window.